learn using anotation, learn send request object and validate input in spring boot
