+++ src/main/java/com/shopapp/shopapp/controllers/CategoryController.java
@@ -1,0 +1,47 @@
+package com.shopapp.shopapp.controllers;
+
+import com.shopapp.shopapp.dtos.CategoryDTO;
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("api/v1/categories")
+//@Validated
+public class CategoryController {
+    //    Hiển thị tất cả category
+    @GetMapping("") //http://localhost:8088/api/v1/categories?page=1&limit=10
+    public ResponseEntity<String> getAllCategories(@RequestParam("page") int page, @RequestParam("limit") int limit) {
+        return ResponseEntity.ok(String.format("getAllCategories, page = %d, limit = %d", page, limit));
+    }
+
+    @PostMapping()
+//    Nếu tham số truyền vào là 1 object thì sao => Data transfer object = request object
+    public ResponseEntity<?> insertCategories(@Valid @RequestBody CategoryDTO categoryDTO, BindingResult result) {
+        if (result.hasErrors()) {
+            List<String> errorMessage = result.getFieldErrors()
+                                                .stream()
+                                                .map(FieldError::getDefaultMessage)
+                                                 .toList();
+            return ResponseEntity.badRequest().body(errorMessage);
+        }
+        return ResponseEntity.ok("This is insert categories" + categoryDTO);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<String> updateCategories(@PathVariable Long id) {
+        return ResponseEntity.ok("This is update categories with id = " + id);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteCategories(@PathVariable Long id) {
+        return ResponseEntity.ok("This is delete categories with id = " + id);
+    }
+
+
+}
