+++ src/main/java/com/shopapp/shopapp/dtos/CategoryDTO.java
@@ -1,0 +1,14 @@
+package com.shopapp.shopapp.dtos;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.*;
+
+@Data //toString
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class CategoryDTO {
+    @NotEmpty(message = "Category's name cannot be empty")
+    private String name;
+}
